// src/enum/products.enum.ts
export enum ProductCategory {
  SMARTPHONES_TABLETS = 'Smartphones & Tablets',
  LAPTOPS_COMPUTERS = 'Laptops & Computers',
  AUDIO_HEADPHONES = 'Audio & Headphones',
  GAMING_CONSOLES = 'Gaming & Consoles',
  CAMERAS_PHOTOGRAPHY = 'Cameras & Photography',
  HOME_APPLIANCES = 'Home Appliances',
  TV_ENTERTAINMENT = 'TV & Entertainment',
  ACCESSORIES = 'Accessories',
  SMART_HOME = 'Smart Home',
  WEARABLES = 'Wearables',
}